backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
litellm==1.30.0
python-multipart==0.0.6
pydantic==2.5.3
python-dotenv==1.0.0
Pillow==10.2.0
google-generativeai==0.8.3
slowapi==0.1.9
pytest==8.0.0
pytest-asyncio==0.23.0
httpx==0.27.0
